backend/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.2.1
django-jsonfield==1.4.0
django-mssql-backend==2.8.1
django-versatileimagefield==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gower==0.0.5
joblib==0.16.0
numpy==1.19.5
pandas==1.1.5
Pillow==7.2.0
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.7.3
python-magic==0.4.18
pytz==2020.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
xlrd==1.2.0
requests==2.24.0
SQLAlchemy==1.3.19
lxml==4.5.2
django-saml2-auth==2.2.1
django-sslserver==0.22
Jinja2==2.11.2
pdfkit==0.6.1
django-redis==4.12.1
python-dotenv==0.19.0
statsmodels==0.12.2
XlsxWriter==3.0.1
azure-msi-mssql-django==1.5
django-easy-audit==1.3.0
openpyxl==3.0.9